Code readability
Splitted Interpreter class into partials.
